parse_expr from scratch
It nows works 100000% better

Operation:
 - If one var ('x'), then just return it
 - If two vars ('x + y'), then create a math expression
 - If three vars ('x + y * z'), put braces according to order
   of precedence ('x + (y * z)')
 - Treat 'x + (y * z)' as two variables ('x' and '(y * z)')
   - Try to parse both vars as expressions.

It also turns out the GCC version of prime.c is much faster if an int
is used instead of a long (6 sec vs 2 sec).

Stack overflows by allocating way too big objects on the stack have also
been fixed by using a malloc followed by a realloc instead.

It is also possible to write directly to stdout or read from stdin for
easier debugging.

The '%' operator now compiles to OP_REM while '%%' compiles to OP_MOD
instead.

There are less temporary variables lingering around due to heavier
(forced) use of the 'istemp' argument.

There are no silly 'snprintf(expr, sizeof expr, "%s == 0", v)'
statements anymore. Instead, LINE_IF inv parameter is set directly
depending on what is needed.

A (now working*) fast version of the prime test has been added, which is
ridiculously fast compared to the naive version (13 vs 0.2 seconds for i
from 3 to 200001)

* The fact that it didn't work spawned the effort to rewrite parse_expr
from scratch.

+++ benchmark/prime-int.c
@@ -9,7 +9,7 @@ int main() {
 		i += 2;
 		if (i == 200001)
 			return 0;
-		for (long j = 2; j < i / 2; j++) {
+		for (int j = 2; j < i / 2; j++) {
 			if (i % j == 0)
 				goto nope;
 		}
